# imgproc: optimise local cost computation in IntelligentScissorsMB::buildMap