When reading the code I noticed a `TODO(opt)` comment, this pull request is for that.

https://github.com/opencv/opencv/blob/4.5.5/modules/imgproc/src/intelligent_scissors.cpp#L628

```
float cost = cost_q + local_cost(q, r);  // TODO(opt): compute partially until cost < cost_r
```

### Notes

* https://github.com/opencv/opencv/pull/21941 and this pull request change the same non-test files. However, the changes themselves don't overlap i.e. there should be no merge conflicts.

* Based on https://github.com/opencv/opencv/wiki/Branches#which-branch-should-i-target-my-pr this pull request here could be for 3.4 instead of 4.x branch. However, it is for 4.x since 3.4 doesn't have the intelligent scissors code.

### Test run snippet

```
$ export OPENCV_TEST_DATA_PATH=../opencv_extra/testdata

$./bin/opencv_test_imgproc --gtest_filter="Imgproc_IntelligentScissorsMB.color/*"

...
[----------] 4 tests from Imgproc_IntelligentScissorsMB
[ RUN      ] Imgproc_IntelligentScissorsMB.color/0, where GetParam() = (true, 1)
[       OK ] Imgproc_IntelligentScissorsMB.color/0 (66 ms)
[ RUN      ] Imgproc_IntelligentScissorsMB.color/1, where GetParam() = (true, 100)
[       OK ] Imgproc_IntelligentScissorsMB.color/1 (5541 ms)
[ RUN      ] Imgproc_IntelligentScissorsMB.color/2, where GetParam() = (false, 1)
[       OK ] Imgproc_IntelligentScissorsMB.color/2 (58 ms)
[ RUN      ] Imgproc_IntelligentScissorsMB.color/3, where GetParam() = (false, 100)
[       OK ] Imgproc_IntelligentScissorsMB.color/3 (5418 ms)
[----------] 4 tests from Imgproc_IntelligentScissorsMB (11083 ms total)
...
```

### Pull Request Readiness Checklist

See details at https://github.com/opencv/opencv/wiki/How_to_contribute#making-a-good-pull-request

- [X] I agree to contribute to the project under Apache 2 License.
- [X] To the best of my knowledge, the proposed patch is not based on a code under GPL or another license that is incompatible with OpenCV
- [X] The PR is proposed to the proper branch
- [ ] There is a reference to the original bug report and related work
- [X] There is accuracy test, performance test and test data in opencv_extra repository, if applicable
      Patch to opencv_extra has the same branch name.
- [ ] The feature is well documented and sample code can be built with the project CMake