@cpoerschke 2022-05-21
Here are some more notes on the optimisation.

The `local_cost(p, q)` is computed as the sum of three terms

```
weight_non_edge_compute * non_edge_feature.at<uchar>(q) +
weight_gradient_direction * fD +
weight_gradient_magnitude * fG
```

and the optimisation uses the non-negativity of the terms and the fact that we don't need the exact cost if the cost exceeds the threshold. E.g. if `term1` increases the cost above the threshold then also adding the non-negative terms `term2` and `term3` cannot decrease the cost to below the threshold and so we can omit computing and adding `term2` and `term3`.

Supporting links w.r.t. non-negativity of the terms:

* `weight_non_edge_compute * non_edge_feature.at<uchar>(q)`
  * https://github.com/opencv/opencv/blob/4.5.5/modules/imgproc/src/intelligent_scissors.cpp#L180
  * https://github.com/opencv/opencv/blob/4.5.5/modules/imgproc/src/intelligent_scissors.cpp#L344
  * `non_edge_feature` via `uchar` type

* `weight_gradient_direction * fD`
  * https://github.com/opencv/opencv/blob/4.5.5/modules/imgproc/src/intelligent_scissors.cpp#L181
  * https://github.com/opencv/opencv/blob/4.5.5/modules/imgproc/src/intelligent_scissors.cpp#L536

* `weight_gradient_magnitude * fG`
  * https://github.com/opencv/opencv/blob/4.5.5/modules/imgproc/src/intelligent_scissors.cpp#L182
  * https://github.com/opencv/opencv/blob/4.5.5/modules/imgproc/src/intelligent_scissors.cpp#L502
  * https://github.com/opencv/opencv/blob/4.5.5/modules/imgproc/src/intelligent_scissors.cpp#L316
  * https://github.com/opencv/opencv/blob/4.5.5/modules/imgproc/src/intelligent_scissors.cpp#L473
  * https://github.com/opencv/opencv/blob/4.5.5/modules/imgproc/src/intelligent_scissors.cpp#L477
  * https://github.com/opencv/opencv/blob/4.5.5/modules/imgproc/include/opencv2/imgproc/segmentation.hpp#L101