@cpoerschke 2022-06-07
Thanks for approving the CI run for this pull request!

Reviewing the _"PR:4.x W10 / BuildAndTest (pull_request) Cancelled after 360m — BuildAndTest"_ outcome, this was

```
Run RET=$(git ls-remote --heads "https://github.com/cpoerschke/opencv_extra" "4.x-intelligent-scissors-optimisation") || true
Error: The operation was canceled.
```

and apparently because (at the time) there was no https://github.com/cpoerschke/opencv_extra fork in existence. Basic git documentation searching found no `git ls-remote` options or other `git` commands that would obviously handle this scenario and so I've just gone ahead and created a fork of the repo. Hope that helps? Thank you.