@cpoerschke 2023-05-23
Hi @asmorkalov - thanks for the review!

> * If I understand correctly, ... Why do you need special flag to switch behaviour?

Your understanding is correct. The special flag (during development) helps compare performance and accuracy via the parameterised tests but yes beyond that there is technically no need for the flag to remain. Sometimes flags can help preserve existing behaviour for APIs for backwards compatibility between versions or so but that is not the case here I think.

> * Could you modify some (not all) tests to check contour values, but not only contour size? It allows to check item 1.

Have just added a commit that modifies (so far only) one test to check ca. 5+5% of the contour values, at the beginning and the end of the contour.