@asmorkalov 2023-05-24
I propose 2 things:
- Dump full contour to file with FileStorage (yaml or yaml.gz) and check full contour in several tests or all tests. You need to create PR to opencv_extra repo with the same branch name. If the branches match they are testsed together by CI.
- Remove flag and stay with optimized branch. I do not see reason to presume original behavior, if they are equivalent. Test with full contour check ensures it.