@cpoerschke 2023-05-25
> I propose 2 things:
> 
> * Dump full contour to file with FileStorage (yaml or yaml.gz) and check full contour in several tests or all tests. You need to create PR to opencv_extra repo with the same branch name. If the branches match they are testsed together by CI.
> ...

#23688 opened for the first thing - .xml files seemed more common than .yaml files and so I went with that for now. https://github.com/opencv/opencv_extra/pull/1065 is the matching PR for the opencv_extra repo.