@cpoerschke 2023-05-26
> I propose 2 things:
> 
> * ...
> * Remove flag and stay with optimized branch. I do not see reason to presume original behavior, if they are equivalent. Test with full contour check ensures it.

https://github.com/opencv/opencv/pull/21959/commits/a01344ecd94435c5f61fce6f166e8f2b26525edd commit added for the second thing.