@asmorkalov 2023-05-27
I propose to restore perf test. We can use summary.py and other scripts to compare performance numbers for different versions, e.g. gefore and after patch.