@asmorkalov 2023-06-07
@cpoerschke Thanks a lot for the contribution! I merged accuracy and performance tests. Could you rebase the PR with optimization on top of current 4.x and squash commits?