@cpoerschke 2023-06-07
> @cpoerschke Thanks a lot for the contribution! I merged accuracy and performance tests. Could you rebase the PR with optimization on top of current 4.x and squash commits?

Thanks @asmorkalov for the review inputs! I've rebased and squashed the commits.

Also ran the performance test on the commit before and the commit itself - unless I'm misreading the numbers the optimisation actually makes things slower now. Would like to look into that more thoroughly, perhaps I missed something when removing the switch flag? The accuracy tests pass.