@asmorkalov 2023-06-08
I see optimization effect, but the mentioned function is not bootle neck:
```
python3 ../opencv-master/modules/ts/misc/summary.py ./IntelligentScissorsMB-4.x.xml ./IntelligentScissorsMB-opt.xml 

Geometric mean (ms)

             Name of Test              IntelligentScissorsMB-4.x IntelligentScissorsMB-opt IntelligentScissorsMB-opt
                                                                                                      vs            
                                                                                           IntelligentScissorsMB-4.x
                                                                                                  (x-factor)        
buildMap::TestIntelligentScissorsMB::0         7995.334                  7819.149                    1.02           
buildMap::TestIntelligentScissorsMB::1         7988.302                  7816.922                    1.02  
```